[#312] update file structure
