+++ objects/invitations/github_invitations.rb
