+++ objects/jobs/job_commiterrors.rb
@@ -18,39 +18,34 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-require 'aws-sdk-s3'
-require 'nokogiri'
-require_relative '../version'
+require_relative '../truncated'
 
 #
-# S3 storage.
+# Job that posts exceptions as commit messages.
+# API: http://octokit.github.io/octokit.rb/method_list.html
 #
-class S3
-  def initialize(ocket, bucket, region, key, secret)
-    @object = Aws::S3::Resource.new(
-      region: region,
-      credentials: Aws::Credentials.new(key, secret)
-    ).bucket(bucket).object(ocket)
+class JobCommitErrors
+  def initialize(name, github, commit, job)
+    @name = name
+    @github = github
+    @commit = commit
+    @job = job
   end
 
-  def load
-    Nokogiri::XML(
-      if @object.exists?
-        data = @object.get.body
-        puts "S3 #{data.size} from #{@object.bucket_name}/#{@object.key}"
-        data
-      else
-        puts "Empty puzzles for #{@object.bucket_name}/#{@object.key}"
-        '<puzzles xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-xsi:noNamespaceSchemaLocation="http://www.0pdd.com/puzzles.xsd"/>'
-      end
+  def proceed
+    @job.proceed
+  rescue Exception => e
+    done = @github.create_commit_comment(
+      @name, @commit,
+      "I wasn't able to retrieve PDD puzzles from the code base and \
+submit them to GitHub. If you \
+think that it's a bug on our side, please submit it to \
+[yegor256/0pdd](https://github.com/yegor256/0pdd/issues):\n\n\
+> #{Truncated.new(e.message.gsub(/\s/, ' '), 300)}\n\n
+Please, copy and paste this stack trace to GitHub:\n\n
+```\n#{e.class.name}\n#{e.message}\n#{e.backtrace.join("\n")}\n```"
     )
-  end
-
-  def save(xml)
-    data = xml.to_s
-    @object.put(body: data)
-    puts "S3 #{data.size} to #{@object.bucket_name}/#{@object.key} \
-(#{xml.xpath('//puzzle').size} puzzles)"
+    puts "Comment posted about an error: #{done['html_url']}"
+    raise e
   end
 end
