+++ objects/jobs/job_detached.rb
