+++ objects/storage/safe_storage.rb
