+++ objects/storage/sync_storage.rb
