+++ objects/storage/upgraded_storage.rb
