+++ test/test_credentials.rb
@@ -24,8 +24,8 @@ require 'yaml'
 require 'tmpdir'
 require 'aws-sdk-dynamodb'
 require_relative 'test__helper'
-require_relative '../objects/s3'
-require_relative '../objects/github_tickets'
+require_relative '../objects/storage/s3'
+require_relative '../objects/tickets/github_tickets'
 require_relative '../objects/log'
 require_relative '../objects/git_repo'
