+++ test/test_logged_tickets.rb
@@ -24,7 +24,7 @@ require 'yaml'
 require_relative 'test__helper'
 require_relative 'fake_log'
 require_relative 'fake_tickets'
-require_relative '../objects/logged_tickets'
+require_relative '../objects/tickets/logged_tickets'
 
 # LoggedTickets test.
 # Author:: Yegor Bugayenko (yegor256@gmail.com)
